[proxy/redis] impr: use redis_auth_type to switch between auth types
This adds `redis_auth_type` to the config with default value of "irsa". Not specifying it will enforce the `regional_redis_client` to be configured with IRSA redis (as it's done now).
If "plain" is specified, then the regional client is condifigured with `redis_notifications`, consuming username:password auth from URI. We plan to do that for Azure.

Configuring `regional_redis_client` is required now, there is no opt-out from configuring it.